Revert "React stopwatch component extra feature"
