+++ react-stopwatch-component/package.json
@@ -1,6 +1,7 @@
 {
   "name": "react-stopwatch-component",
   "version": "1.0.0",
+  "description": "",
   "main": "index.js",
   "scripts": {
     "build": "webpack",
@@ -11,15 +12,14 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.13.10",
-    "@babel/plugin-transform-react-jsx": "^7.13.12",
+    "@babel/plugin-transform-react-jsx": "^7.12.17",
     "babel-loader": "^8.2.2",
-    "webpack": "^5.27.2",
+    "webpack": "^5.27.0",
     "webpack-cli": "^4.5.0"
-  },
-  "description": ""
+  }
 }
